refactor(core): refactor device management API
I'm going to make a Config class that is a bag of fields.
JSON will parse to that AND this will parse to it.
Based on the compatible field, we will convert to the concrete Configuration and create the concrete device directly.
So basically ConcreteDevice::Configuration knows how to validate itself.
