+++ core/core/interactive_test/hostobjs/config.cpp
@@ -1,0 +1,6 @@
+#include "./vocab.hpp"
+#include "core/interactive_test/dict.hpp"
+#include "core/interactive_test/vocab/core_words.hpp"
+#include "fmt/format.h"
+
+void register_config_words(Interpreter *p) {}
