+++ core/core/interactive_test/hostobjs/vocab.cpp
@@ -1,0 +1,3 @@
+#include "./vocab.hpp"
+
+void register_devicemgmt_words(Interpreter *interp) { register_value_words(interp); }
